Remove external reference (and baggage) of System.Runtime.InteropServices.RuntimeInformation.dll. This work also removes the need for the Microsoft.Win32.Registry NuGet package
