+++ src/Datadog.Trace/FrameworkDescription.cs
@@ -8,7 +8,7 @@ using Microsoft.Win32;
 
 namespace Datadog.Trace
 {
-    internal class FrameworkDescription
+    internal partial class FrameworkDescription
     {
         private static readonly Vendors.Serilog.ILogger Log = DatadogLogging.GetLogger(typeof(FrameworkDescription));
 
@@ -56,154 +56,12 @@ namespace Datadog.Trace
 
         public string ProcessArchitecture { get; }
 
-        public static FrameworkDescription Create()
-        {
-            var assemblyName = RootAssembly.GetName();
-
-            if (string.Equals(assemblyName.Name, "mscorlib", StringComparison.OrdinalIgnoreCase))
-            {
-                // .NET Framework
-                return new FrameworkDescription(
-                    ".NET Framework",
-                    GetNetFrameworkVersion() ?? "unknown",
-                    "Windows",
-                    Environment.Is64BitOperatingSystem ? "x64" : "x86",
-                    Environment.Is64BitProcess ? "x64" : "x86");
-            }
-
-            // .NET Core
-            return CreateFromRuntimeInformation();
-        }
-
         public override string ToString()
         {
             // examples:
             // .NET Framework 4.8 x86 on Windows x64
             // .NET Core 3.0.0 x64 on Linux x64
             return $"{Name} {ProductVersion} {ProcessArchitecture} on {OSPlatform} {OSArchitecture}";
-        }
-
-        private static FrameworkDescription CreateFromRuntimeInformation()
-        {
-            string frameworkName = null;
-            string osPlatform = null;
-
-            try
-            {
-                // RuntimeInformation.FrameworkDescription returns a string like ".NET Framework 4.7.2" or ".NET Core 2.1",
-                // we want to return everything before the last space
-                string frameworkDescription = RuntimeInformation.FrameworkDescription;
-                int index = frameworkDescription.LastIndexOf(' ');
-                frameworkName = frameworkDescription.Substring(0, index).Trim();
-            }
-            catch (Exception e)
-            {
-                Log.SafeLogError(e, "Error getting framework name from RuntimeInformation");
-            }
-
-            if (RuntimeInformation.IsOSPlatform(System.Runtime.InteropServices.OSPlatform.Windows))
-            {
-                osPlatform = "Windows";
-            }
-            else if (RuntimeInformation.IsOSPlatform(System.Runtime.InteropServices.OSPlatform.Linux))
-            {
-                osPlatform = "Linux";
-            }
-            else if (RuntimeInformation.IsOSPlatform(System.Runtime.InteropServices.OSPlatform.OSX))
-            {
-                osPlatform = "MacOS";
-            }
-
-            return new FrameworkDescription(
-                frameworkName ?? "unknown",
-                GetNetCoreVersion() ?? "unknown",
-                osPlatform ?? "unknown",
-                RuntimeInformation.OSArchitecture.ToString().ToLowerInvariant(),
-                RuntimeInformation.ProcessArchitecture.ToString().ToLowerInvariant());
-        }
-
-        private static string GetNetFrameworkVersion()
-        {
-            string productVersion = null;
-
-            try
-            {
-                object registryValue;
-
-                using (var baseKey = RegistryKey.OpenBaseKey(RegistryHive.LocalMachine, RegistryView.Default))
-                using (var subKey = baseKey.OpenSubKey(@"SOFTWARE\Microsoft\NET Framework Setup\NDP\v4\Full\"))
-                {
-                    registryValue = subKey?.GetValue("Release");
-                }
-
-                if (registryValue is int release)
-                {
-                    // find the known version on the list with the largest release number
-                    // that is lower than or equal to the release number in the Windows Registry
-                    productVersion = DotNetFrameworkVersionMapping.FirstOrDefault(t => release >= t.Item1)?.Item2;
-                }
-            }
-            catch (Exception e)
-            {
-                Log.SafeLogError(e, "Error getting .NET Framework version from Windows Registry");
-            }
-
-            if (productVersion == null)
-            {
-                // if we fail to extract version from assembly path,
-                // fall back to the [AssemblyInformationalVersion] or [AssemblyFileVersion]
-                productVersion = GetVersionFromAssemblyAttributes();
-            }
-
-            return productVersion;
-        }
-
-        private static string GetNetCoreVersion()
-        {
-            string productVersion = null;
-
-            if (Environment.Version.Major == 3 || Environment.Version.Major >= 5)
-            {
-                // Environment.Version returns "4.x" in .NET Core 2.x,
-                // but it is correct since .NET Core 3.0.0
-                productVersion = Environment.Version.ToString();
-            }
-
-            if (productVersion == null)
-            {
-                try
-                {
-                    // try to get product version from assembly path
-                    Match match = Regex.Match(
-                        RootAssembly.CodeBase,
-                        @"/[^/]*microsoft\.netcore\.app/(\d+\.\d+\.\d+[^/]*)/",
-                        RegexOptions.IgnoreCase);
-
-                    if (match.Success && match.Groups.Count > 0 && match.Groups[1].Success)
-                    {
-                        productVersion = match.Groups[1].Value;
-                    }
-                }
-                catch (Exception e)
-                {
-                    Log.SafeLogError(e, "Error getting .NET Core version from assembly path");
-                }
-            }
-
-            if (productVersion == null)
-            {
-                // if we fail to extract version from assembly path,
-                // fall back to the [AssemblyInformationalVersion] or [AssemblyFileVersion]
-                productVersion = GetVersionFromAssemblyAttributes();
-            }
-
-            if (productVersion == null)
-            {
-                // at this point, everything else has failed (this is probably the same as [AssemblyFileVersion] above)
-                productVersion = Environment.Version.ToString();
-            }
-
-            return productVersion;
         }
 
         private static string GetVersionFromAssemblyAttributes()
